+++ test/test_GUNW.py
@@ -45,7 +45,8 @@ def test_GUNW():
 
     ## check the CRS and affine are written correctly
     epsg      = 4326
-    transform = (0.1, 0.0, -119.35, 0, -0.1, 35.05)
+    transform = (0.1, 0.0, -119.85, 0, -0.1, 35.55)
+
     group = f'science/grids/corrections/external/troposphere/{WM}/reference'
     for v in 'troposphereWet troposphereHydrostatic'.split():
         ds = rio.open(f'netcdf:{updated_GUNW}:{group}/{v}')
@@ -65,6 +66,6 @@ def test_GUNW():
 
     # Clean up files
     shutil.rmtree(SCENARIO_DIR)
-    os.remove('GUNW_20200130-20200124.yaml')
+    os.remove('GUNW_20200130-20200124_135156.yaml')
     [os.remove(f) for f in glob.glob(f'{WM}*')]
     return