+++ test/test_datelist.py
@@ -10,7 +10,11 @@ from test import TEST_DIR, WM
 
 def test_datelist():
     SCENARIO_DIR = os.path.join(TEST_DIR, 'datelist')
+
+    if os.path.exists(SCENARIO_DIR):
+        shutil.rmtree(SCENARIO_DIR) 
     os.makedirs(SCENARIO_DIR, exist_ok=False)
+
     dates = ['20200124', '20200130']
 
     dct_group = {