+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.4.3]
++ Force lat/lon/hgt to float32 so that they line up correctly in stitching
++ Add two stage buffer; 
+    + first pad user bounding box such that a 3D cube is generated that at min covers user area of interest.
+    + then if ray tracing is used, pad the downloaded model in look direction. Assumes look angle is fixed increases with latitude.
+       
++ Update and convert user given AOI to weather model projection (except for HRRR)
 + Clean up error messagse, skip date if temporal interpolation fails
 + Update valid range for ERA5 (current date - 3 months) & ERA5T
 + Temporal interpolation of delays if the requested datetime is more than _THRESHOLD_SECONDS away from the closest weather model available time and `interpolate_time = True` (default behavior)